[feat] 바텀 시트 공통 ui 분리
